74 bytes free in monitor now. Wohoo! I factored the common code out of "measures" and the 'v', 't' and 'l' commands and put it in common/measure.s43 as 3 routines called cellV, temp and linkV.
Changed ADC10 to use MCLK and divide it by 2 itself. That way we're free to do what we like with SMCLK for the timer. Not that there's anything we need to change there AFAIK.

Renamed measAndCal to measAndScale since the values it returns for temp and linkV are not fully calibrated.

Changed 'c' for commsState to 'r' for RxState and made it return ticks since last Rx. Monitor only.